requirements.txt
numpy==1.22.3
scikit-learn==1.0.2
pymongo==4.1.1
pytest~=8.2.2
mongomock~=4.1.2
aniso8601==9.0.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==2.0.12
click==8.1.7
exceptiongroup==1.2.1
Flask==2.0.3
flask-restx==1.3.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mongomock==4.1.2
numpy==1.22.3
packaging==24.1
pluggy==1.5.0
pymongo==4.1.1
pytest==8.2.2
pytz==2024.1
referencing==0.35.1
requests==2.27.1
rpds-py==0.18.1
scikit-learn==1.0.2
scipy==1.11.4
sentinels==1.0.0
six==1.16.0
threadpoolctl==3.5.0
tomli==2.0.1
urllib3==1.26.18
Werkzeug==2.2.2

python-dotenv~=1.0.1